refactor(agent): split provider internals
